requirements.txt
discord.py==2.2.2
fluent.runtime==0.4.0